J cthedevelop3r/clean the code

+++ api/controllers/dashboard.js
@@ -1,7 +1,3 @@
-const User = require("../models/user");
-const Catway = require("../models/catway");
-const Reservation = require("../models/reservation");
-const bcrypt = require("bcrypt");
 const userService = require("../services/user");
 const dashboardService = require("../services/dashboard");
 
@@ -11,17 +7,36 @@ async function createUser(req, res) {
 
     // Vérifier que tous les champs sont remplis
     if (!name || !email || !password) {
-      return res.status(400).send("Tous les champs sont requis.");
+      throw new Error("Tous les champs doivent être remplis.");
     }
 
     // Appeler le service pour créer l'utilisateur
     await userService.createUser(name, email, password);
 
+    req.session.successCreateUser = "Utilisateur créé avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successCreateUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
     // Redirection après création réussie
     res.redirect("/dashboard");
   } catch (error) {
-    console.error(error);
-    res.status(500).send(error.message || "Erreur serveur.");
+    console.error("Erreur lors de la création de l'utilisateur :", error.message);
+    
+    req.session.errorCreateUser = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorCreateUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
+    res.redirect("/dashboard");
   }
 }
 
@@ -30,40 +45,77 @@ async function updateUser(req, res) {
     const { userId, name, email } = req.body;
 
     // Vérifier que tous les champs nécessaires sont présents
-    if (!name || !email) {
-      return res.status(400).send("Tous les champs sont requis.");
+    if (!userId || !name || !email) {
+      throw new Error("Tous les champs doivent être remplis.");
     }
 
     // Appeler le service pour mettre à jour l'utilisateur
     await userService.updateUser(userId, name, email);
 
+    req.session.successUpdateUser = "Utilisateur modifié avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successUpdateUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
     res.redirect("/dashboard");
   } catch (error) {
-    console.error(error);
-    res.status(500).send(error.message || "Erreur serveur.");
+    console.error("Erreur lors de la mise à jour de l'utilisateur :", error.message);
+    
+    req.session.errorUpdateUser = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorUpdateUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
+    res.redirect("/dashboard");
   }
 }
 
 async function deleteUser(req, res) {
-  console.log("🗑️ Requête reçue pour suppression :", req.body);
-
   try {
     const { userId } = req.body;
 
-    // Vérifier que l'ID est bien fourni
+    // Vérifier que tous les champs sont remplis
     if (!userId) {
-      return res.status(400).send("ID de l'utilisateur requis.");
+      throw new Error("L'ID est requis.")
     }
 
     // Appeler le service pour supprimer l'utilisateur
     await userService.deleteUser(userId);
 
+    req.session.successDeleteUser = "Utilisateur supprimé avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successDeleteUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
     res.redirect("/dashboard");
   } catch (error) {
-    console.error(error);
-    res.status(500).send(error.message || "Erreur serveur.");
+    console.error("Erreur lors de la suppression de l'utilisateur :", error);
+
+    req.session.errorDeleteUser = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorDeleteUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
+    res.redirect("/dashboard");
   }
 }
+
 
 async function createCatway(req, res) {
   try {
@@ -71,16 +123,35 @@ async function createCatway(req, res) {
 
     // Vérifier que les champs nécessaires sont fournis
     if (!type || !catwayState) {
-      return res.status(400).send("Type et état du catway sont requis.");
+      throw new Error("Le type du catway et la description de l'état du catway sont requis.")
     }
 
     // Appeler le service pour créer le catway
     await dashboardService.createCatway(type, catwayState);
 
+    req.session.successCreateCatway = "Catway créé avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successDeleteUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
     res.redirect("/dashboard");
-  } catch (err) {
-    console.error("Erreur lors de la création du catway :", err);
-    res.status(500).send("Erreur serveur lors de la création du catway");
+  } catch (error) {
+    console.error("Erreur lors de la création du catway :", error);
+
+    req.session.errorCreateCatway = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorCreateCatway = null;
+        req.session.save();
+      }, 10000);
+    });  
+
+    res.redirect("/dashboard");
   }
 }
 
@@ -88,8 +159,8 @@ async function getNextCatwayNumber(req, res) {
   try {
     const nextCatwayNumber = await dashboardService.getNextCatwayNumber();
     res.json({ nextCatwayNumber }); // Retourne le numéro dans une réponse JSON
-  } catch (err) {
-    console.error("Erreur lors du calcul du numéro du catway :", err);
+  } catch (error) {
+    console.error("Erreur serveur : ", error)
     res.status(500).send("Erreur serveur");
   }
 }
@@ -99,15 +170,32 @@ async function updateCatwayState(req, res) {
     const { catwayId, catwayState } = req.body;
 
     if (!catwayId || !catwayState) {
-      return res.status(400).json({ error: "ID Catway et État sont requis." });
+      throw new Error("Tous les champs sont requis.");
     }
 
     await dashboardService.updateCatwayState(catwayId, catwayState);
 
+    req.session.successUpdateCatway = "L'opération est un succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successUpdateUser = null;
+        req.session.save();
+      }, 10000);
+    });
+
     res.redirect("/dashboard");
   } catch (error) {
-    console.error("Erreur lors de la mise à jour de l'état du Catway :", error.message);
-    return res.status(500).json({ error: "Erreur serveur lors de la mise à jour." });
+    req.session.errorUpdateCatway = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorUpdateCatway = null;
+        req.session.save();
+      }, 10000);
+    });  
+
+    res.redirect("/dashboard");
   }
 }
 
@@ -115,16 +203,38 @@ async function deleteCatway(req, res) {
   try {
     const { catwayNumber } = req.body;
 
+    const numCatway = Number(catwayNumber);
+    if (isNaN(numCatway)) {
+      throw new Error("Numéro de catway invalide.");
+    }
+
     if (!catwayNumber) {
-      return res.status(400).json({ error: "Numéro du catway requis." });
+      throw new Error("Numéro du catway requis.")
     }
 
     await dashboardService.deleteCatway(catwayNumber);
 
+    req.session.successDeleteCatway = "Catway supprimé avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successDeleteCatway = null;
+        req.session.save();
+      }, 10000);
+    });
+
     res.redirect("/dashboard");
   } catch (error) {
-    console.error("Erreur suppression catway :", error.message);
-    res.status(500).json({ error: "Erreur serveur." });
+    req.session.errorDeleteCatway = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorDeleteCatway = null;
+        req.session.save();
+      }, 10000);
+    });  
+
+    res.redirect("/dashboard");
   }
 }
 
@@ -135,7 +245,6 @@ async function getCatwayDetails(req, res) {
     const catwayDetails = await dashboardService.getCatwayDetails(catwayNumber);
     res.json(catwayDetails);
   } catch (error) {
-    console.error("Erreur récupération catway :", error.message);
     res.status(404).json({ error: "Catway non trouvé" });
   }
 }
@@ -144,19 +253,54 @@ async function saveReservation(req, res) {
   try {
     await dashboardService.createReservation(req.body);
     res.redirect("/dashboard");
+
+    req.session.successSaveReservation = "Réservation enregistrée avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successSaveReservation = null;
+        req.session.save();
+      }, 10000);
+    });
   } catch (error) {
-    console.error("Erreur lors de l'enregistrement :", error.message);
-    res.status(400).send(error.message);
+    req.session.errorSaveReservation = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorSaveReservation = null;
+        req.session.save();
+      }, 10000);
+    });  
+
+    res.redirect("/dashboard");
   }
 }
 
 async function deleteReservation(req, res) {
   try {
     await dashboardService.deleteReservation(req.body.reservationId);
+
+    req.session.successDeleteReservation = "Réservation supprimée avec succès !";
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.successDeleteReservation = null;
+        req.session.save();
+      }, 10000);
+    });
+
     res.redirect("/dashboard");
   } catch (error) {
-    console.error("Erreur suppression réservation :", error.message);
-    res.status(400).json({ error: error.message });
+    req.session.errorDeleteReservation = error.message;
+
+    req.session.save(() => {
+      setTimeout(() => {
+        req.session.errorDeleteReservation = null;
+        req.session.save();
+      }, 10000);
+    });  
+
+    res.redirect("/dashboard");
   }
 }
 
@@ -165,7 +309,6 @@ async function displayReservationDetails(req, res) {
     const reservationDetails = await dashboardService.getReservationDetails(req.params.reservationId);
     res.json(reservationDetails);
   } catch (error) {
-    console.error("Erreur récupération réservation :", error.message);
     res.status(400).json({ error: error.message });
   }
 }
